+++ packages/ts-newskit/src/components/slices/shared/customTextBlock.tsx
@@ -1,28 +1,23 @@
 import React, { ReactNode } from 'react';
-import { TextBlock } from 'newskit';
-import { Wrapper } from '../shared-styles';
+import { InlineTextBlock } from '../shared-styles';
 
 export const CustomTextBlock = ({
   text,
   stylePreset,
-  icon,
-  alignFlex = true
+  icon
 }: {
   text?: string;
   stylePreset?: string;
   icon?: ReactNode;
-  alignFlex?: boolean;
 }) => {
   return (
-    <TextBlock
+    <InlineTextBlock
       typographyPreset="utilityLabel005"
       stylePreset={stylePreset ? stylePreset : 'inkBrand010'}
       as="span"
     >
-      <Wrapper as="span" alignFlex={alignFlex}>
-        {icon}
-        {text}
-      </Wrapper>
-    </TextBlock>
+      {icon}
+      {text}
+    </InlineTextBlock>
   );
 };